HS-0: fix clicking bug in adventure mode
It wasn't possible to left-click a block when in adventure mode. This is now fixed by using the PlayerAnimationEvent

Note that this can cause the tile to be automatically clicked when unflagging a flagged tile (This was seen on Spigot, Paper seems to not show this behavior). It is not possible to fix this, due to not being able to distinguish between left and right clicks with the PlayerAnimationEvent.